[NEW] Payment mixin
