+++ l10n_br_fiscal/models/document_fiscal_mixin_methods.py
@@ -1,5 +1,4 @@
 # Copyright (C) 2019  Renato Lima - Akretion <renato.lima@akretion.com.br>
-# Copyright (C) 2020  Luis Felipe Mileo - KMEE <mileo@kmee.com.br>
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
 from odoo import api, models, _
@@ -33,73 +32,3 @@ class FiscalDocumentMixinMethods(models.AbstractModel):
         if self.fiscal_operation_id:
             self.operation_name = self.fiscal_operation_id.name
             self.comment_ids = self.fiscal_operation_id.comment_ids
-
-    def check_financial(self):
-        for record in self:
-            if not record.env.context.get('action_document_confirm'):
-                continue
-            elif record.amount_missing_payment_value > 0:
-                if not record.payment_term_id:
-                    raise UserError(
-                        _("O Valor dos lançamentos financeiros é "
-                          "menor que o valor da nota."),
-                    )
-                else:
-                    record.generate_financial()
-
-    def generate_financial(self):
-        for record in self:
-            if record.payment_term_id and self.company_id and self.currency_id:
-                record.financial_ids.unlink()
-                record.fiscal_payment_ids.unlink()
-                vals = {
-                    'payment_term_id': self.payment_term_id.id,
-                    'amount': self.amount_missing_payment_value,
-                    'currency_id': self.currency_id.id,
-                    'company_id': self.company_id.id,
-                }
-                vals.update(self.fiscal_payment_ids._compute_payment_vals(
-                    payment_term_id=self.payment_term_id,
-                    currency_id=self.currency_id,
-                    company_id=self.company_id,
-                    amount=self.amount_missing_payment_value, date=self.date)
-                )
-                self.fiscal_payment_ids = self.fiscal_payment_ids.new(vals)
-                for line in self.fiscal_payment_ids.mapped('line_ids'):
-                    line.document_id = self
-
-            elif record.fiscal_payment_ids:
-                record.financial_ids.unlink()
-                record.fiscal_payment_ids.unlink()
-
-    @api.onchange("fiscal_payment_ids", "payment_term_id")
-    def _onchange_fiscal_payment_ids(self):
-        financial_ids = []
-
-        for payment in self.fiscal_payment_ids:
-            for line in payment.line_ids:
-                financial_ids.append(line.id)
-        self.financial_ids = [(6, 0, financial_ids)]
-
-    # @api.onchange("payment_term_id", "company_id", "currency_id",
-    #               "amount_missing_payment_value", "date")
-    # def _onchange_payment_term_id(self):
-    #     if (self.payment_term_id and self.company_id and
-    #             self.currency_id):
-    #
-    #         self.financial_ids.unlink()
-    #
-    #         vals = {
-    #             'payment_term_id': self.payment_term_id.id,
-    #             'amount': self.amount_missing_payment_value,
-    #             'currency_id': self.currency_id.id,
-    #             'company_id': self.company_id.id,
-    #          }
-    #         vals.update(self.fiscal_payment_ids._compute_payment_vals(
-    #             payment_term_id=self.payment_term_id, currency_id=self.currency_id,
-    #             company_id=self.company_id,
-    #             amount=self.amount_missing_payment_value, date=self.date)
-    #         )
-    #         self.fiscal_payment_ids = self.fiscal_payment_ids.new(vals)
-    #         for line in self.fiscal_payment_ids.mapped('line_ids'):
-    #             line.document_id = self
